Remove some warnings on Visual Studio .Net

+++ VTK/IO/vtkDataReader.cxx
@@ -145,7 +145,7 @@ void vtkDataReader::SetInputString(const char *in)
 { 
   if (in != NULL)
     {
-    this->SetInputString(in, strlen(in));
+    this->SetInputString(in, static_cast<int>(strlen(in)));
     }
   else
     {
@@ -2141,7 +2141,7 @@ void vtkDataReader::DecodeArrayName(char *resname, const char* name)
   //strcpy(resname, name);
   ostrstream str;
   int cc = 0, ch;
-  int len = strlen(name);
+  int len = static_cast<int>(strlen(name));
   char buffer[10] = "0x";
   while(name[cc]) 
     {
